+++ docs/changelog.rst
@@ -1,6 +1,34 @@
 Changelog
 ~~~~~~~~~
 
+Unreleased
+----------
+
+* New feature: ``add_qubit`` to add fresh qubits at specified positions in an ``MPS``.
+* New feature: added an option to ``measure`` to toggle destructive measurement on/off. Currently only supported for ``MPS``.
+* New feature: a seed can now be provided to ``Config`` objects, providing reproducibility across ``StructuredState`` simulations.
+* New feature: ``apply_unitary`` both for ``MPS`` and ``TTN`` to apply an arbitrary unitary matrix, rather than a ``pytket.Command``.
+* New feature: ``apply_qubit_relabelling`` both for ``MPS`` and ``TTN`` to change the name of their qubits. This is now used within ``simulate`` to take into account the action of implicit SWAPs in pytket circuits (no additional SWAP gates are applied).
+
+0.6.1 (April 2024)
+------------------
+
+* When using ``simulate`` with ``TTNxGate`` algorithm, the initial partition is obtained using NetworkX instead of KaHyPar by default. This makes setup easier and means that ``TTNxGate`` can now be used when installing from PyPI. KaHyPar can still be used if ``use_kahypar`` from ``Config`` is set to True.
+* Updated pytket version requirement to 1.27.
+
+0.6.0 (April 2024)
+------------------
+
+* **New feature**: Tree Tensor Network (TTN) simulator, supporting both fixed ``chi`` and ``truncation_fidelity``. Calculation of single amplitudes is supported by ``get_amplitude`` and inner products by ``vdot``. Measurement and postselection are not yet supported.
+* **New API**: both ``MPS`` and ``TTN`` share a common interface: ``StructuredState``. Import paths have changed, multiple classes have been renamed: ``ConfigMPS`` is now ``Config``, ``ContractionAlg`` is now ``SimulationAlgorithm``. Documentation has been updated accordingly.
+
+* Canonicalisation of MPS is now always applied before a two-qubit gate. We found that this tends to reduce runtime due to canonicalisation decreasing virtual bond dimension.
+* Two-qubit gates are now decomposed (SVD) before applying them to remove null singular values (e.g. in ``XXPhase`` gates).
+* Fixed a bug on copying an ``MPS`` if ``truncation_fidelity`` was set.
+* Fixed a bug on ``CuTensorNetHandle`` that would prevent it from working when the device set was different from the default one (``dev=0``) and when using ``cuTensorNet>=2.3.0``.
+* Fixed a bug on ``TensorNetwork`` due to unsupported ``Create`` operation.
+* Updated pytket version requirement to 1.26.
+
 0.5.4 (January 2024)
 --------------------